Add files via upload
The is a jupyter notebook application used to process an image, detect if there are faces in the image and if there are return the facial image.
Use:
status,uncropped_files_list, to_small_list, no_face_list =   crop_image(source_dir, dest_dir, mode, min_img_size)
where:
source_dir is the full path to a directory containing image files to be processed
dest_dir is the full path to the directory where the cropped images will be stored
mode is an integer. If mode is = 1 then for each image processed the application finds the LARGEST facial image in the image and stores that single cropped image to the dest_dir. If mode is NOT 1 then the application finds ALL facial images in the image being processed and stores ALL these cropped facial images in the destination directory
min_img_size is an integer denoting the minimum size (both height and width) a cropped facial image must exceed in order to be saved to the dest_dir
The application returns the following:
status - a boolean. Set to True if no errors were found in the input parameters and at least one good facial image was stored in the dest_dir
uncropped_files_list is a list denoting the input files that were not cropped
to_small_list is a list of input file paths for which all the faces found in the image file were smaller than the min_img_size and were not saved to the dest_dir
no_face_list is a list of input file paths for which no face was found in the image

NOTE: This application usesCV2, and  MTCNN which requires Tensorflow. Both MTCNN, Tensorflow and CV2 must be installed in the operating environment
MTCNN while very good is NOT flawless. Always check the images in the dest_dir to ensure they are properly cropped