Add files via upload
This function splits the files in the s_dir into train, test, and validation files stored in the dest_dir.
s_dir is the full path to the directory containing the files to be split
dest_dir is the full path to the destination directory. If it does not exist it is created.
train_size is a float between 0.0 and 1.0 indicating the percentage of file to be allocated as training files
test_size is a float between 0.0 and 1.0 indicating the percentage of file to be allocated as test files
In the dest_dir three sub directories 'train', 'test' and 'valid' are createdand used to store the training files,
test files and validation files.
If these sub directories already exist they are check for existing content.If content is found a notice is printed
to that effect. The user is then prompted to enter 'D' to delete the content, 'Q' to terminate program execution
or 'C' to continue. If 'C' is selected the content is not removed however files may be over written if any existing
files have the same file name as the new files being added to the sub directory.
Note if the test, train and valid directories exist and have content, and the user elects 'c' to continue
sub directories and files from the s_dir are appended to the content of the test, train and valid subdirectories
in the dest_dir
This function utlilizes tqdm and sklearn which must be installed in your working environment