Add files via upload
This function operates on image files and renames files numerically and can also resize the images
If you download images files from an internet search you will notice the files often have very long
and un-weildy names often with strange characters in the name. Also these files may come in a
multitude of image formats as denotd by their extensions. After I download these images I find it
convenient to rename the files and generally to convert them all to a common image format.
This function achieves that purpose. It will nename the files in a numerical order starting with a
number you specificy.
You specify a new file extension to converted all files to the image format specified by the extension.
Alternatively you can specify parameter new_ext='same' to preserve the original extension.
You can also elect to risize all the images by specifying the parameters height and width.
File are nenamed in a numerical sequence starting with the value assigned to parameter snum.
Note- files are rename using "zeros padding" which preserves the order that files are processed
in for python functions. Unlike windows, files in most python functions read files alphanumerically.
For example is you have files in a directory with names 1.jpg, 2.jpg, ... 9.jpg, 10,jpg the order in which
the will be processed is not the numeric order but the alphabetic order. Thus they will be
processed in the order 1.jpg, 10.jpg, 2.jpg etc. The use of "zeros' padding will prevent this.
With zero padding the files are renamed 01.jpg, 02.jpg .. 09.jpg, 10.jpg so the are processed in
the order you would expect.
Use:
1- place the images you want processed in a single directory. I call it the source_dir. place ONLY image files in the directory.
NOTE this function use the cv2 module so file types are limited to having extensions 'jpg', 'jpe', 'jpeg', 'png', 'bmp', 'tiff', 'jfif' .
2 create a python environment with the module cv2 (OpenCV) and tqdm installed
3 The function call is of the form rename (source_dir, snum=1, new_ext='same', height=None, width=None) where:
dir is a string denoting the full path to the directory containing the image files
snum is an integer (default value=1) that defines the starting number for the numerical renaming of the files
new_ext is a string (default= 'same) denoting the file format all image files will be converted to.
If set to 'same' the original extensions will be preserved
height is an integer specify the new height of all images - defualts to None - meaning no resizing of the images
width is an integer specify the new width of all images - defualts to None - meaning no resizing of the images
Processing:
1 The function first checks the input parameters and prints out a message if errors are detected
2 Next the files in dir are checked to verify the file extensions are in the list of acceptable extensions
as defined in the list above. If files are found with invalid extensions, a sub directory called
"bad_ext" is created within the source_dir and the file with the improper extension is removed from
source_dir and stored in the sub directory. A bad_list is created and the filenameshaving improper
extensions are stored in that list and provided as part of the functions return tupple.
3 Next each of the remaining files in dir are processed sequentially. CV2 is used to read in the image file.
If an error is encounter the image file is corrupted and can not be processed.The file name is added
to the bad_list and the file is removed from the source_dir and placed into the "bad_ext subdirectory
4 Now the source_dir directory only contains image files that can be processed. These files are then
renamed sequentially start from snum and if parameters height and width are specified the image
is resized. If parameter new_ext is not set to 'same' then the files are written back to the source_dir
with the new extension you specified
returns:
The function returns a tupple of the form (status, bad_list) where:
status is a boolean. It is set to true if at least one image file was processed to completetion.
Otherwise it is set to False
bad_list is a list of file names. If files in dir were found to have extensions not in the list above,
or if an image file could not be read by cv2 the file name is included in the list.
These files will be stored in the sub directory bad ext in the source_dir¶