Add files via upload
This function operates on a directory- s_dir assuming to have sub directories that contain
image files. For each subdirectory the list of files is processed to determine if the image files have allowable
extensions and that the image file is not corrupted (ie it can be read by cv2). If a file has an invalid extension
or is corrupted it's file path is appended to a list called bad_list. The file is then removed from the sub directory
and placed in a sub directory of the s_dir called bad_files.
the list of allowable extensions is provided by the good_exts parameter
a typical list of extensions is ['jpg', 'png', 'bmp','tiff','jpeg', 'gif'] and should all be lower case¶