Add files via upload
This program searches through a directory and detects images that are duplicates.
First it goes through all the files in the directory that are image files and extracts the files extension.
It then determines if the file extension is in a list of file extension types that can be processed.
If the extension is in the list, the file's path is added to a list called f_list for further processing.
If not in the extension list the file's path is appended to list named bad_lis and is not processedt.
Then for the files in the f_list it reads in the file in as gray scale, and resizes it into a vector of length
hash_length +1. It then performs a difference hash on the vector. The file name along with the hash
is stored in a dictionary hash_dict, with the key=file name and the value = the hash.
Once the dictionary is made for all files in the f_list the dictionary is then processed to find if any files have
hash values that are with 'distance' of each other. These files are then considered duplicates of each other.
A new dictionary dup_dict is then created. The dictionary key is a file name. The value is a list of all file names
that are duplicates of the key file name. The user is given the option to delete the duplicate files.
if duplicate files are deleted the user is given the option of nenaming the files in a numerically sequential
fashion starting from a specified starting number. Note- files are renamed sequentially using
'zero's padding' which preserves the processing order of files by python functions
Note this function requires tqdm and cv2 be installed in your environment¶