Add files via upload
This function downloads image files from BING and stores them into a directory.
The form of the call to the function is
download(output_dir,query,snum, new_ext, limit=100, verbose=0, timeout=10) where
output_dir is the full path to the directory where the downloaded files will be stored
query is a string defining the subject of the files to download - for example query='bird images'
limit is an integer specifying the maximum number of files to download
verbose is an integer. If verbose=0 details of the download process are not printed out
timeout is an integer specifying the max time in seconds to wait for a query response
When downloading images from BING, the file names are typically long and often unconventional
and can be difficult to handle if you process the downloaded images
this function includes the ability to rename the downloaded images in a numerical sequence
starting with the numerical value of the integer snum. If snum=0 no renaming of the downloaded
files takes place. Files are renamed in numerical sequence with "zeros padding" The number of
leading zeros in the name is a function of the number of files stored in the output_dir.
If less than 10 no leading zeros are present, if less than 100, 1 leading 0 is appended. If more
1000 are present two leading zeros are present. Zeros padding preserves the file order
processing in typical python functions
If snum is not equal to 0 the files are renumbered and you also have the option to modify
the image format to that specified by the string next_ext. If new_ext='same' the original file
formats are preserved. If new_ext='jpg" for example all downloaded images are converted
to a jpg format.
Note: see the import cell below to determine what modules need to be installed¶