Add files via upload
This function acts as a generator to produce batches of numpy arrays and labels read from a director
to use first instantiate the class with
mygen=data_gen(s_dir, batch_size, data_shape, shuffle) .flow()where
s_dir is the full path to the directory contain two subdirectories labeled 'data' and 'labels'
the data sub directory hold the arrays with extension .npy
the label directory holds the arrays of integer label values with extension .npy
batch size is an integer of the number of values the generator will return
data_shape is a tuple defining the shape of the arrays in the data sub directory
shuffle is a binary. If true the data for each batch is randomly selected. If set to False
the data is selected in sequential order as present in the directories
,flow causes the generator to read in the data and produce the batch output of the generator
The outpus of the generator are arrays, labels, array_paths, label_paths
arrays are the input arrays of the shape(batch_size, data_shape)
labels are input arrays of the form shape(batch_size,)
array_paths is an ordered list associated with the input arrays specifying the array name
label_paths is an ordered list associated with the input labels specifying the array name¶