Add files via upload
This function acts as a generator to produce batches of numpy arrays and labels read from a director
to use first instantiate the class with
mygen=data_gen(s_dir, batch_size, data_shape, shuffle, num_of_classes,categorical) .flow()where
s_dir is the full path to the directory contain two subdirectories labeled 'data' and 'labels'
the data sub directory hold the arrays with extension .npy
the label directory holds the arrays of integer label values with extension .npy
batch size is an integer of the number of values the generator will return
data_shape is a tuple defining the shape of the arrays in the data sub directory
shuffle is a binary. If true the data for each batch is randomly selected. If set to False
the data is selected in sequential order as present in the directories
num_of_classes is an integer specifying the number of classes, that is the maximum
value of a label. Only used when categorical=True
categorical is a binary. If False labels are provided as integers
If categorical=True labels are provided as 1 hot vectors
,flow causes the generator to read in the data and produce the batch output of the generator
The outpus of the generator are arrays, labels, array_paths, label_paths
arrays are the input arrays of the shape(batch_size, data_shape)
labels are input arrays of the labels of the form shape(batch_size,) when categorical=False
and shape(batch_size, num_of_classes if categorical is true
The generator has two attributes which may be useful
.label_path contains a sequentially ordered list of all label file names produced by the generator
.array_path contains a sequentially ordered list of all array file names produced by the generator
array_paths is an ordered list associated with the input arrays specifying the array name
label_paths is an ordered list associated with the input labels specifying the array name