Add files via upload
This function renames files in a directory and if specied will resize the images
rename (source_dir, snum, new_ext, resize) where:
source_dir is the full path to the directory containing the image files to be processed
snum is an integer. If set to 0 the file names will not be changes but the files can be still be converted
to a new image format and also be resized
if snum is a non 0 integer the files are renamed in numerical sequence starting with the value of snum.
note file are renumber with "zeros" padding so that files are processedby python functions based on
their numerical order.
new_ext is a string specifying the new image format the files should be converted to.
note only these extensions are allowed 'jpg', 'jpe', 'jpeg', 'png', 'bmp', 'tiff'
if next_ext='same' the files will be kept in their original format
resize specifies the new size for the images. If resize='same' the images are new resized
if not set to 'same' resize must be a tuple of the form (height,width) where height and width
are integers.
To use this function cv2 and tqdm must be installed in your working environment
Note only files of the type listed under extensions above can be processed.
if a file in the source_dir is not of the type that can be processed by cv2.imread, an exception
occurs and a message is printed saying the file has been deleted from the directory.¶