Add files via upload
This function reads in images from a directory and predicts the images class using a trained model
predict_images(path_to_images, path_to_model, path_to_class_dict, img_shape, preprocess_function,show_images,reverse_function=None ) where
path_to_images is the full path to the directory containing the images
path_to_model is the full path to the the location of the saved trained model
path_to_class_dict is the full path to a text file which is the text version of a dictionary of the form {class_index, class_name}
images shape is a tuple of the form (image_height, image_width, bands) for example (224,224,3) This tuple must be the
the same as the image size the model was trained on
preprocess_function is a function you must define that was used in the original model to preprocess the images. For
example if image piixels were scalled to be between -1 to +1 see example of use for definition of the function
show_images is a boolean. If set to true a plot is made with each image shown with its predicted label
reverse_function is a function you must define to rescale the pixel back to a range of 0 to 1 so they can be displayed
only needed if show_images=True